% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputVerification.R
\name{input_check}
\alias{input_check}
\title{input_check}
\usage{
input_check(X, X_hat)
}
\arguments{
\item{X}{\itemize{
\item Target matrix (hxk) for k variables and forecasting horizon h
}}

\item{X_hat}{\itemize{
\item Predicted matrix (hxk) for k variables and forecasting horizon h
}}
}
\value{
\itemize{
\item NULL
}
}
\description{
Verifies that the input values to the error measure functions have the correct type and shapes
}
