% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariateErrorMeasures.R
\name{MMASE}
\alias{MMASE}
\title{MMASE - Multivariate Mean Absolute Scaled Error}
\usage{
MMASE(X, X_hat)
}
\arguments{
\item{X}{\itemize{
\item Target matrix (hxk) for k variables and forecasting horizon h
}}

\item{X_hat}{\itemize{
\item Predicted matrix (hxk) for k variables and forecasting horizon h
}}
}
\value{
List containing:
\itemize{
\item{\code{values_vec}: }{Vector of the individual MASEs}
\item{\code{mean}: }{Mean of the vector of the individual MASEs}
}
}
\description{
Computes the multivariate MASE as a mean of the MASEs of the individual time series
}
\examples{
X <- matrix(rnorm(100),10,10)
X_hat <- matrix(rnorm(100,mean=1),10,10)
results_list <- MMASE(X,X_hat)
results_list$values_vec # For individual MASEs vector
results_list$mean # For the mean of the individual MASEs vector
}
