% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariateErrorMeasures.R
\name{MMSE}
\alias{MMSE}
\title{MMSE - Multivariate Mean Squared Error}
\usage{
MMSE(X, X_hat)
}
\arguments{
\item{X}{\itemize{
\item Target matrix (hxk) for k variables and forecasting horizon h
}}

\item{X_hat}{\itemize{
\item Predicted matrix (hxk) for k variables and forecasting horizon h
}}
}
\value{
List containing:
\itemize{
\item{\code{values_vec}: }{Vector of the individual MSEs}
\item{\code{mean}: }{Mean of the vector of the individual MSEs}
}
}
\description{
Computes the multivariate MMSE as a mean of the MSEs of the individual time series
}
\examples{
X <- matrix(rnorm(100),10,10)
X_hat <- matrix(rnorm(100,mean=1),10,10)
results_list <- MMSE(X,X_hat)
results_list$values_vec # For individual MSEs vector
results_list$mean # For the mean of the individual MSEs
}
